requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
python-dotenv==1.0.0
openai==1.2.3
youtube-transcript-api==0.6.1
google-api-python-client==2.108.0
python-multipart==0.0.6
httpx==0.25.1
